+++ app/src/main/java/com/opera/app/fragments/HomeFragment.java
@@ -16,15 +16,19 @@ import com.opera.app.R;
 import com.opera.app.activities.MainActivity;
 import com.opera.app.activities.PreLoginActivity;
 import com.opera.app.activities.SearchEventActivity;
+import com.opera.app.constants.AppConstants;
 import com.opera.app.controller.MainController;
 import com.opera.app.dagger.Api;
 import com.opera.app.database.events.EventGenresDB;
 import com.opera.app.database.events.EventListingDB;
 import com.opera.app.listadapters.CoverFlowAdapter;
 import com.opera.app.listadapters.WhatsOnPagerAdapter;
+import com.opera.app.listener.MarkFavouriteInterface;
 import com.opera.app.listener.TaskComplete;
 import com.opera.app.pojo.events.eventlisiting.AllEvents;
 import com.opera.app.pojo.events.eventlisiting.Events;
+import com.opera.app.pojo.favouriteandsettings.Favourite;
+import com.opera.app.pojo.favouriteandsettings.FavouriteAndSettingsResponseMain;
 import com.opera.app.preferences.SessionManager;
 import com.opera.app.utils.LanguageManager;
 
@@ -48,8 +52,8 @@ public class HomeFragment extends BaseFragment {
     private EventGenresDB mEventGenresDB;
     private ArrayList<Events> mHighlightedEvents = new ArrayList<>();
     private ArrayList<Events> mWhatsEvents = new ArrayList<>();
+    private ArrayList<Favourite> arrFavouriteDataOfLoggedInUser = new ArrayList<>();
     private SessionManager manager;
-
     private WhatsOnPagerAdapter mWhatsOnPagerAdapter;
     //    private AdapterEvent mAdapterEvent;
     private ArrayList<Events> mEventAllData = new ArrayList<>();
@@ -88,7 +92,17 @@ public class HomeFragment extends BaseFragment {
 
         InitView(view);
         GetCurrentEvents();
+
+        if(manager.isUserLoggedIn()){
+            GetUserSettings();
+        }
+
         return view;
+    }
+
+    private void GetUserSettings() {
+        MainController controller = new MainController(mActivity);
+        controller.getUpdatedSettings(taskComplete, api);
     }
 
     private void GetCurrentEvents() {
@@ -117,7 +131,6 @@ public class HomeFragment extends BaseFragment {
         mViewpagerWhatsOnShows.setOnPageChangeListener(new ViewPager.OnPageChangeListener() {
             @Override
             public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
-                Log.e("onPageScrolled", position + "");
                 if (position == 0) {
                     mViewpagerWhatsOnShows.setPadding(0, 0, 70, 0);
                 } else if (mWhatsEvents.size() - 1 == position) {
@@ -140,30 +153,40 @@ public class HomeFragment extends BaseFragment {
     private TaskComplete taskComplete = new TaskComplete() {
         @Override
         public void onTaskFinished(Response response, String mRequestKey) {
-            AllEvents mEventDataPojo = (AllEvents) response.body();
+            if (mRequestKey.equalsIgnoreCase(AppConstants.GETUSERSETTINGS.GETUSERSETTINGS)) {
+                FavouriteAndSettingsResponseMain mFavouriteAndSettingsResponseMain = (FavouriteAndSettingsResponseMain) response.body();
 
-            try {
-                if (mEventDataPojo.getStatus().equalsIgnoreCase("success")) {
-                    mEventListingDB.open();
-
-                    //Calling this function for guest user to take his favourite events
-                    if (!manager.isUserLoggedIn()) {
-                        mEventAllData = new ArrayList<>();
-                        mEventAllData = mEventListingDB.fetchEventsWithFavouriteForGuest();
-                    }
-
-                    mEventListingDB.deleteCompleteTable(EventListingDB.TABLE_EVENT_LISTING);
-                    mEventListingDB.insertOtherEvents(mActivity, mEventDataPojo.getEvents(), mEventAllData);
-
-                    mEventGenresDB.open();
-                    mEventGenresDB.deleteCompleteTable(EventGenresDB.TABLE_GENRES_LISTING);
-                    mEventGenresDB.insertEventsGenres(mEventDataPojo.getGenreList());
-
-                    fetchDataFromDB();
-
+                Log.e("GetUserSetting response","true");
+                if (mFavouriteAndSettingsResponseMain.getStatus().equalsIgnoreCase("success")) {
+                    arrFavouriteDataOfLoggedInUser.addAll(mFavouriteAndSettingsResponseMain.getData().getFavourite());
+                    UpdateFavouriteData();
+                    GetWhatsOnEvents();
                 }
-            } catch (Exception e) {
-                e.printStackTrace();
+            } else {
+                AllEvents mEventDataPojo = (AllEvents) response.body();
+                try {
+                    if (mEventDataPojo.getStatus().equalsIgnoreCase("success")) {
+                        mEventListingDB.open();
+
+                        //Calling this function for guest user to take his favourite events
+                        if (!manager.isUserLoggedIn()) {
+                            mEventAllData = new ArrayList<>();
+                            mEventAllData = mEventListingDB.fetchEventsWithFavouriteForGuest();
+                        }
+
+                        mEventListingDB.deleteCompleteTable(EventListingDB.TABLE_EVENT_LISTING);
+                        mEventListingDB.insertOtherEvents(mActivity, mEventDataPojo.getEvents(), mEventAllData,arrFavouriteDataOfLoggedInUser);
+
+                        mEventGenresDB.open();
+                        mEventGenresDB.deleteCompleteTable(EventGenresDB.TABLE_GENRES_LISTING);
+                        mEventGenresDB.insertEventsGenres(mEventDataPojo.getGenreList());
+
+                        fetchDataFromDB();
+
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
             }
         }
 
@@ -180,9 +203,22 @@ public class HomeFragment extends BaseFragment {
     };
 
     private void fetchDataFromDB() {
+        Log.e("Event Listing response","true");
+        UpdateFavouriteData();
         GetWhatsOnEvents();
         GetHighlightedEvents();
         mEventListingDB.close();
+        arrFavouriteDataOfLoggedInUser = new ArrayList<>();
+    }
+
+    private void UpdateFavouriteData() {
+        if (arrFavouriteDataOfLoggedInUser.size() > 0) {
+            mEventListingDB.open();
+            for (int i = 0; i < arrFavouriteDataOfLoggedInUser.size(); i++) {
+                mEventListingDB.UpdateFavouriteData(arrFavouriteDataOfLoggedInUser.get(i).getFavouriteId().toUpperCase(), arrFavouriteDataOfLoggedInUser.get(i).getIsFavourite());
+            }
+            mEventListingDB.close();
+        }
     }
 
     private void GetWhatsOnEvents() {
